test: test product and user features

+++ src/modules/products/services/delete-category-service.ts
@@ -6,10 +6,16 @@ export default class DeleteCategoryService {
   constructor (
     private readonly categoryRepository: ICategoryRepository) {}
 
-  public async execute (id: any): Promise<boolean> {
+  public async execute (id?: any): Promise<boolean> {
     if (!id) {
       throw new AppError('Missing param: id')
     }
+
+    const category = await this.categoryRepository.findById(id)
+    if (!category) {
+      throw new AppError('Category not found')
+    }
+
     return await this.categoryRepository.delete(id)
   }
 }
